+++ jkit-sql/src/main/java/com/alianga/jkit/sql/SqlParser.java
@@ -956,7 +956,7 @@ public final class SqlParser {
         return raw;
     }
 
-    private boolean isStmtSeparator() {
+    boolean isStmtSeparator() {
         if (is(SqlTokenType.GO)) {
             return true;
         }